requirements.txt
pyserial==3.5
keyboard==0.13.5